refactor(agents): remove dead helpers and re-export shims (A2)
Delete test-only economics, model_diversity, ensemble dispatch, and
reviewer_merge modules; drop legacy stdout parsers and CLI trace_jsonl
shim. Move finding_key to review/ as the sole production consumer.
