+++ src/mergecraft/agents/codex.py
@@ -16,7 +16,6 @@ from loguru import logger
 
 from mergecraft.agents.codex_broker import (
     OPENAI_API_KEY_ENV,
-    CodexBrokeredRun,
     active_broker_handle,
     add_broker_openai_config,
     begin_broker_session,
@@ -29,7 +28,6 @@ from mergecraft.agents.codex_broker import (
 from mergecraft.agents.codex_stream import (
     CODEX_MODEL_REASONING_EFFORT,
     codex_stream_event_handler,
-    parse_codex_payload,
 )
 from mergecraft.agents.openai_compatible_gateways import (
     CUSTOM_PROVIDER_API_KEY_ENV,
@@ -43,7 +41,6 @@ from mergecraft.agents.reviewer import REVIEWER_AGENT_NAME, REVIEWER_SYSTEM_PROM
 from mergecraft.agents.shared import (
     AgentResult,
     AgentRunContext,
-    AgentUsage,
     agent,
     payload_shell_mode,
     spawn_agent_cli,
@@ -252,10 +249,6 @@ def _save_codex_writeback_state(*, auth_path: Path, auth_json: str) -> None:
 
 def _has_openai_api_key() -> bool:
     return bool(os.environ.get(OPENAI_API_KEY_ENV, "").strip())
-
-
-def _codex_subscription_auth_usable(raw: str) -> bool:
-    return subscription_auth_usable(raw)
 
 
 def _setup_codex_auth(
@@ -634,42 +627,6 @@ def _build_env(ctx: AgentRunContext) -> dict[str, str]:
     return env
 
 
-def _parse_codex_stdout(stdout: str) -> tuple[str, AgentUsage | None]:
-    text = stdout.strip()
-    if not text:
-        return "", None
-
-    try:
-        data = json.loads(text)
-        if isinstance(data, dict):
-            return parse_codex_payload(data)
-    except json.JSONDecodeError:
-        pass
-
-    usage: AgentUsage | None = None
-    output = text
-    for line in reversed(text.splitlines()):
-        stripped = line.strip()
-        if not stripped:
-            continue
-        try:
-            event = json.loads(stripped)
-        except json.JSONDecodeError:
-            continue
-        if not isinstance(event, dict):
-            continue
-        if event.get("type") == "message" and isinstance(event.get("content"), str):
-            output = str(event["content"])
-        parsed_output, parsed_usage = parse_codex_payload(event)
-        if parsed_output:
-            output = parsed_output
-        if parsed_usage is not None:
-            usage = parsed_usage
-        if event.get("type") in {"turn.completed", "agent-turn-complete", "result"}:
-            break
-    return output, usage
-
-
 def _run_codex_once(
     *,
     cli: str,
@@ -903,17 +860,6 @@ async def _run(ctx: AgentRunContext) -> AgentResult:
     finally:
         stop_broker_session(broker_session)
         set_broker_session(None)
-
-
-def prepare_codex_brokered_run(
-    ctx: AgentRunContext,
-    *,
-    openai_api_key: str = "",
-) -> CodexBrokeredRun:
-    """Start broker, build env, auth stub, and MCP config (plan 18 W3)."""
-    from mergecraft.agents import codex_broker
-
-    return codex_broker.prepare_codex_brokered_run(ctx, openai_api_key=openai_api_key)
 
 
 codex = agent(name="codex", install=_install, run=_run, build_env=_build_env)
