+++ src/mergecraft/cli/run_cmd.py
@@ -14,11 +14,14 @@ from typing import Any
 import typer
 from rich.table import Table
 
-from mergecraft.cli import trace_jsonl
 from mergecraft.cli.consoles import err_console as console
 from mergecraft.cli.exits import CLI_SUCCESS_EXIT_CODE
 from mergecraft.cli.global_surface import emit_cli_json, wants_json_output
-from mergecraft.cli.trace_jsonl import load_trace_jsonl_events, session_ids_in_trace_order
+from mergecraft.tracing.trace_jsonl import (
+    default_trace_dir,
+    load_trace_jsonl_events,
+    session_ids_in_trace_order,
+)
 
 app = typer.Typer(
     name="run",
@@ -76,7 +79,7 @@ def inspect_cmd(
     ),
 ) -> None:
     """Inspect a stored review run (or list known run ids)."""
-    target = trace_dir if trace_dir is not None else trace_jsonl.default_trace_dir()
+    target = trace_dir if trace_dir is not None else default_trace_dir()
     events = _load_events(target)
     sessions = _sessions(events)
     chosen = run_id or (sessions[-1] if sessions else None)
@@ -109,7 +112,7 @@ def diff_cmd(
     ),
 ) -> None:
     """Compare two stored review runs by event kind."""
-    target = trace_dir if trace_dir is not None else trace_jsonl.default_trace_dir()
+    target = trace_dir if trace_dir is not None else default_trace_dir()
     events = _load_events(target)
     sessions = _sessions(events)
     left = run_a
